Bump version to 0.1.10; add tempfile dependency; apply clippy lints and simplify code patterns
Add `tempfile` crate for safer temp file handling in XLSX conversion. Replace manual temp file cleanup with `NamedTempFile` that auto-deletes on drop. Implement `Default` trait for types with `new()` constructors. Simplify conditional patterns using `let` chains. Remove redundant `.trim()` from `normalize_whitespace`. Replace `Self::default()` with direct struct initialization for zero-sized types. Use

+++ src/timeseries.rs
@@ -231,7 +231,7 @@ impl TimeSeriesProcessor {
                     chrono::NaiveTime::from_hms_opt(0, 0, 0).unwrap_or(chrono::NaiveTime::MIN),
                 );
                 let duration_since_epoch = timestamp.signed_duration_since(epoch);
-                let rounded_duration = (duration_since_epoch.num_seconds() as i64
+                let rounded_duration = (duration_since_epoch.num_seconds()
                     / duration.num_seconds())
                     * duration.num_seconds();
                 epoch + Duration::seconds(rounded_duration)
